Move GoogleAI types to single file
