--- packages/vertexai/src/types/googleAI/requests.ts
@@ -1,29 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { Content, Part } from '../content';
-import { GenerationConfig, Tool } from '../requests';
-
-export interface GoogleAICountTokensRequest {
-  generateContentRequest: {
-    model: string; // 'models/model-name'
-    contents: Content[];
-    systemInstruction?: string | Part | Content;
-    tools?: Tool[];
-    generationConfig?: GenerationConfig;
-  };
-}
