--- packages/vertexai/src/types/googleAI/responses.ts
@@ -1,46 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { Content } from '../content';
-import { FinishReason } from '../enums';
-import {
-  Citation,
-  GroundingMetadata,
-  PromptFeedback,
-  SafetyRating,
-  UsageMetadata
-} from '../responses';
-
-export interface GoogleAIGenerateContentResponse {
-  candidates?: GoogleAIGenerateContentCandidate[];
-  promptFeedback?: PromptFeedback;
-  usageMetadata?: UsageMetadata;
-}
-
-export interface GoogleAIGenerateContentCandidate {
-  index: number;
-  content: Content;
-  finishReason?: FinishReason;
-  finishMessage?: string;
-  safetyRatings?: SafetyRating[];
-  citationMetadata?: GoogleAICitationMetadata;
-  groundingMetadata?: GroundingMetadata;
-}
-
-export interface GoogleAICitationMetadata {
-  citationSources: Citation[]; // Maps to `citations`
-}
