+++ packages/vertexai/src/types/index.ts
@@ -22,3 +22,4 @@ export * from './responses';
 export * from './error';
 export * from './schema';
 export * from './imagen';
+export * from './googleai';
